requirements.txt
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
Deprecated==1.2.13
Faker==15.3.4
idna==3.4
pycparser==2.21
PyGithub==1.57
PyJWT==2.6.0
PyNaCl==1.5.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7
requests==2.28.1
schedule==1.1.0
six==1.16.0
urllib3==1.26.13
wrapt==1.14.1
